@@ -6,8 +6,9 @@ import (
 	"github.com/echocat/kubor/command"
 	"github.com/echocat/kubor/common"
 	"github.com/echocat/kubor/kubernetes"
-	"github.com/echocat/kubor/log"
 	"github.com/echocat/kubor/model"
+	"github.com/echocat/slf4g/native"
+	"github.com/echocat/slf4g/native/facade/value"
 	"os"
 	"runtime"
 	"time"
@@ -47,14 +48,25 @@ func main() {
 
 	app := kingpin.New("kubor", "Safely bringing repositories using templating and charting inside CI/CD pipelines to Kubernetes.").
 		ErrorWriter(os.Stderr).
-		UsageWriter(os.Stderr).
-		PreAction(func(_ *kingpin.ParseContext) error {
-			return log.DefaultLogger.Init()
-		})
+		UsageWriter(os.Stderr)
+
+	lv := value.NewProvider(native.DefaultProvider)
+
+	app.Flag("logLevel", "Specifies the minimum required log level.").
+		Envar("KUBOR_LOG_LEVEL").
+		Default(lv.Level.String()).
+		SetValue(&lv.Level)
+	app.Flag("logFormat", "Specifies format output (text or json).").
+		Envar("KUBOR_LOG_FORMAT").
+		Default(lv.Consumer.Formatter.String()).
+		SetValue(&lv.Consumer.Formatter)
+	app.Flag("logColorMode", "Specifies if the output is in colors or not (auto, never or always).").
+		Envar("KUBOR_LOG_COLOR_MODE").
+		Default(lv.Consumer.Formatter.ColorMode.String()).
+		SetValue(lv.Consumer.Formatter.ColorMode)
 
 	pf.ConfigureFlags(app)
 	kubernetes.ConfigureKubeConfigFlags(app)
-	log.DefaultLogger.ConfigureFlags(app)
 
 	if err := common.ConfigureCliCommands("", app, extVersion); err != nil {
 		panic(err)
